[Examples] Fix build time getting progressively slower (#5873)
* Separate needed functions into utils/*.mjs to make them reusable

* examples/rollup.config.js: use separated utils/script-target.mjs

* separate buildTarget

* Refactor buildTarget to accomodate for different input/buildDir

* cleanup

* only build what's needed

---------
