+++ examples/rollup.config.js
@@ -12,6 +12,9 @@ import replace from '@rollup/plugin-replace';
 import resolve from "@rollup/plugin-node-resolve";
 import terser from '@rollup/plugin-terser';
 
+import { buildTarget } from '../utils/rollup-build-target.mjs';
+import { scriptTarget } from '../utils/rollup-script-target.mjs';
+
 /** @typedef {import('rollup').RollupOptions} RollupOptions */
 /** @typedef {import('rollup').Plugin} RollupPlugin */
 
@@ -30,12 +33,8 @@ const staticFiles = [
     { src: './assets', dest: 'dist/static/assets/' },
     { src: '../scripts', dest: 'dist/static/scripts/' },
     { src: '../build/playcanvas.d.ts', dest: 'dist/playcanvas.d.ts' },
-    { src: '../build/playcanvas.js', dest: 'dist/iframe/playcanvas.js' },
-    { src: '../build/playcanvas.dbg.js', dest: 'dist/iframe/playcanvas.dbg.js' },
-    { src: '../build/playcanvas.prf.js', dest: 'dist/iframe/playcanvas.prf.js' },
-    { src: '../build/playcanvas-extras.js', dest: 'dist/iframe/playcanvas-extras.js' },
     { src: './node_modules/@playcanvas/observer/dist/index.js', dest: 'dist/iframe/playcanvas-observer.js' },
-    { src: './node_modules/monaco-editor/min/vs', dest: 'dist/node_modules/monaco-editor/min/vs' },
+    { src: './node_modules/monaco-editor/min/vs', dest: 'dist/node_modules/monaco-editor/min/vs' }
 ];
 
 // ^ = beginning of line
@@ -55,7 +54,7 @@ const regexpImport         =  /^\s*import\s*.+\s*;\s*$/gm;
  *    import './polyfill/OESVertexArrayObject.js';
  *`);
  * @param {string} content - The file content to test.
- * @returns {boolean}
+ * @returns {boolean} Whether content is a module.
  */
 function isModuleWithExternalDependencies(content) {
     const a = regexpExportStarFrom.test(content);
@@ -100,7 +99,7 @@ function timestamp() {
 }
 
 /**
- * @param {import('rollup').Plugin} - The Rollup plugin.
+ * @param {import('rollup').Plugin} plugin - The Rollup plugin.
  * @param {string} src - File or path to watch.
  */
 function watch(plugin, src) {
@@ -179,34 +178,8 @@ const aliasEntries = {
     '@playcanvas/pcui/styles': PCUI_STYLES_PATH
 };
 
-/**
- * Build an ES5 target that rollup is supposed to build.
- *
- * @param {string} name - The name, like `pcx` or `VoxParser`.
- * @param {string} input - The input file, like `extras/index.js`.
- * @param {string} output - The output file, like `dist/iframe/playcanvas-extras.js`.
- * @returns {RollupOptions} One rollup target.
- */
-function scriptTarget(name, input, output) {
-    return {
-        input: input,
-        output: {
-            name: name,
-            file: output,
-            format: 'umd',
-            indent: '\t',
-            globals: { playcanvas: 'pc' }
-        },
-        plugins: [
-            resolve()
-        ],
-        external: ['playcanvas'],
-        cache: false
-    };
-}
-
 /** @type {RollupOptions[]} */
-const builds = [
+const targets = [
     {
         // A debug build is ~2.3MB and a release build ~0.6MB
         input: 'src/app/index.mjs',
@@ -242,4 +215,26 @@ const builds = [
     scriptTarget('pcx', '../extras/index.js', 'dist/iframe/playcanvas-extras.js')
 ];
 
-export default builds;
+// We skip building PlayCanvas ourselves when ENGINE_PATH is given.
+// In that case we have a watcher which copies all necessary files.
+if (ENGINE_PATH === '') {
+    /** @type {RollupOptions|undefined} */
+    let target;
+    if (NODE_ENV === 'production') {
+        // Outputs: dist/iframe/playcanvas.js
+        target = buildTarget('release', 'es5', '../src/index.js', 'dist/iframe');
+    } else if (NODE_ENV === 'development') {
+        // Outputs: dist/iframe/playcanvas.dbg.js
+        target = buildTarget('debug', 'es5', '../src/index.js', 'dist/iframe');
+    } else if (NODE_ENV === 'profiler') {
+        // Outputs: dist/iframe/playcanvas.prf.js
+        target = buildTarget('profiler', 'es5', '../src/index.js', 'dist/iframe');
+    } else {
+        console.warn("NODE_ENV is neither production, development nor profiler.");
+    }
+    if (target) {
+        targets.push(target);
+    }
+}
+
+export default targets;
